Abstract out read buffer; fixed race condition
The read buffer is now abstracted out as the BoundedBuffer class and test
independently. The previously mentioned race condition is fixed in this PTL
based scheme. An added benefit is that this will make it easier when moving
from a fixed number of segments to a dynamic number based on contention.