Use a simpler, faster ring buffer implementation
The previous approach used a PTL style algorithm as a simple hack to replace a
non-strict lossy buffer with a sequential implementation. The original, used
by CLHM, allowed garbage elements to remain due to the cache being used in high
throughput environments. As a more general purpose cache, Caffeine should avoid
retaining garbage entries in the buffer as throughput is unknown.

The PTL workaround was not ideal as it created garbage by placing a ticket in
the array location (a boxed Long). This is replaced with a classic algorithm
that watches the two counters, tries to append if not full, and is garbage
free.