Add Ehcache and Infinispan to benchmark for comparison
Of note is that Infinispan's rewrite to use the backport of JDK8's CHM
decreased performance. Previously they used a Guava-like cache, but
the rewrite uses a lock-free deque. This causes contention as all
threads try to move elements to the tail, resulting in sequential
access.

Ehcache's rewrite to the backport has only minor gains, due to their
heavy use of compute methods resulting in lock contention.