Add buffer variant that uses spacing to avoid false sharing
This version provides more direct control of the cache line behavior
to avoid false sharing, ensuring that the line is used by only one
element. The AtomicReference version may allow 4 references to per
line, so there is a greater chance of conflicts.

On a 4 core machine the performance is equivalent. This should be
tested on a 16 or 32 core machine to see if the change is valuable.