Share common expiration tests.
Added an support for the test to declare it needs either access or
write expiration. The generator filters out combinations where
neither are enabled.

A similar requirement is needed for future reference tests, so
added as well.

This removes a lot of bloat due to duplications with only metadata
changes.