Silently publish the JCache event
This slight change is to ensure that the `ignoreSynchronous()` does not
ignore other events already recorded. However, we want to ignore any
events occuring on the background executor. Since these types of
evictions are not user facing and may be asynchronous, they are not
required to be blocking calls to user-facing threads.