Add Wikipedia request tracing to simulator
10% of the requests over 9/19/2007 to 1/2/2008 are provided by the
wikibench project. Each trace file is 1 hour, consuming 200mb compressed
(1.2GB raw). To reduce the size and simulation time, these files are
preprocessed to replace the URLs with numeric ids and discard unused
columns. This reduces the trace to 20mb compressed (48mb raw).

Due to the size of the files they are not checked in.