Fix slow tests task not running all the slow tests
The slow tests task was incorrectly filtered resulting in it running
only a small subset of the `ReferenceTest` tests. This resulted in a
slight contract difference between Guava and Caffeine. Previously
Caffeine reported `true` for `asMap().containsKey(key)` when the value
had been GC'd but the entry present. Guava checked for this condition
and returned `false. Caffeine now matches Guava and reports `false`.

Some Caffeine only tests relied on this behavior for its assertions.
Those now use a less exact but equally valid approach.

Thankfully this build mistake was small, due to running many of these
tests in the IDE during development. Running the whole test class to
verify a Guava bug fix is what identified this problem.