JCache TCK bug isn't being fixed, change workaround
The TCK is prone to races due to assuming no asynchronous actions may
occur. This results in asynchronous state being intercepted by other
tests due to poor socket handling. Previously the build tried to
isolate this to avoid leaking a fix into the code, but the JSR is not
actively maintained. Since this and other bugs are being ignored,
promote to a change in the implementation.