Optimize 4-bit CountMinSketch
The significant change was to use Unsafe for frequency (4x faster) and
only reset if the increment was successful (33% faster). A small gain
was observed with manual loop unrolling for increment. All these changes
helped in one case, hurt in the other, so its fickle.

The final JMH benchmark numbers are,
 * frequency  533,822,666 ops/s
 * increment   98,869,969 ops/s