Rename EdenQueue to WindowTinyLfu
The new name better describes the policy, whereas the old name was how
I originally stumbled onto that configuration when thinking through
ways to mix policy ideas.

The window calculation was adjusted to ensure that it is not zero for
small caches. The old calculation allowed the gzip trace to exhibit
TinyLfu's poor performance. This calculation matches LIRS, which needs
a similar HIR window to be optimal.