Support varying weights with W-TinyLfu
The weight of an entry may change during its lifetime. The previous
approach of using the node's weight worked fine for a single queue
(LRU), but when using multiple queues a race is visible. The entry
being moved from the eden to the main queue, with a concurrent
update to its weight, could result in the wrong size calculation.
When the update is replayed it adjusts the main queue, leaving the
eden queue at some incorrect size. Over time the eden queue may
either shrink or grow beyound the expected boundary (1%) because
of this race.

The solution is to use store the policy's view of the weight in
addition to the node's. This allows the policy to adjust correctly.
However when an entry is removed an update may be pending (out of
order) and the policy's weight cannot be relied upon. Since the
node has been finalized (dead) at that point, we continue to use
the node's weight instead.

This quirk was something we ran into when first experimenting with
LIRS in CLHM. So we knew the solution this time around, since any
cache that moves entries between queues and supports weights will
have a sizing race. The option is either to abandon multiple queues
(considered for CLHM, done by Infinispan) or handle it by maintaining
a policy-oriented view as well.