Perform replace() in compute block
A similar issue to resurrecting an expired value occurs with weights of
zero. An atomic change of the weight should have to be performed in a
compute block so that the proper fix can be applied. The non-compute
version of replace & put come from CLHM, which didn't have that feature
and was pre-JDK8.

This issue remains for put(), but that has proven harder to solve so this
is going in alone. A fixed put() is almost the same as a compute() via
remap(). That reuse doesn't quite work due to small differences, such as
asserting the negative weight should happen prior to calling the writer.
So the fix may either be a modified copy of remap(), more flags on
remap(), or the current logic transformed to a compute call.

After fixing put, adding the resurrection and zero weight handling to
evictMain() are required. A unit test would be good, if possible.