Rewrite put() to use compute
As discussed in the previous commit, this is needed for resurrecting
an entry about to be evicted but lost the race that set weight=0. This
also is in some ways a simpler version to verify correctness and is
a lighter variant of remap(). The changes do lose some optimizations,
but those probably didn't help much due to writes dominated by other
factors.