Implemented TinyCache as a policy
With dense hash table for TinyCache, change the cache size to 512 so it
is a power of two.
I am looking into it, but the number of replacements from TinyCache is
especially low - as far as I can tell it is a lot lower than any other
policy.  372 evictions on Multi1 trace.

This is very nice news because the addition is by design slower.