RemovalListener not part of primary API
Lambdas with generics on the builder is ugly, as they are inferred as
raw types unless specified. The notification wrapper never evolved to
add more value in Guava, so we can safely remove it here.

This first step is just to discard the wrapper in the public API. The
next step is to consider if it should be removed entirely.