Read performance tuning
This bumps performance from 150M to 180M ops/s on my laptop. It looks
like scheduling is the main culprit and the fast path doesn't aleviate
it as hoped. Removing that logic bumps it to 285M, but isolating code
to remove drops it to 135M. This leads me to think that false sharing
is creating contention and some fields need padding. That might then
get us up into the mid 200s.