Initialize sketch when getting close to the maximum
The initial laziness is to avoid unnecessary bloat for an empty cache.
Afterwards the sketch must grow, but doesn't need to create garbage
for the intermediate transitions. When weighted its unclear what the
maximum number of entries will be. But for the common case, we can
delay until a reasonable starting point. That's either when the
additions reach a threshold or, if sized large and then shrunk, if
the new maximum is within the threshold. Until then the sketch is a
no-op with a size 1, but that state is discarded when growing.