Rewrite of Wikipedia parser to avoid preprocessing step
Instead of preprocessing the input file(s) and assigning a unique id, the
parser processes each line and uses a 128-bit hash to assign an id. The
simulator can now read in multiple trace files as a single sequence. After
adding a few optimizations, the performance is acceptable and the hit rates
are nearly identical.