Review time usage in jcache
Removed redundant calls to get the current time, which is an expensive
native call if performed excessively. This uncovered that the MXBean
was returning times in milliseconds instead of microsecond resolutions.
Its unfortunate that the TCK and spec don't provide a time source to
help validate all usages by a controlled clock.