Optimized the JCache adapter
A simple benchmark showed that, under stress, the cache was performing
only 20M reads/s and 9M/s writes on a laptop. This is a far cry from
the results when benchmarking against the core library.

The reads were dominated by native calls to acquire the current time for
expiration and statistics. Both of these are dynamic configurations, so
they cannot be factored out at construction. Instead smarter handling
is used to avoid unnecessary calls. This boosted it to 135M/s.

The write rate increased to 35M/s after removing unnecessary event
dispatch work and acquiring the current time. The GC churn seems to
account for 2-3M/s, which isn't enough given that writes are not the
common case and the complexity required to reduce it. The copier was
another 2-3M/s, since its enabled, but the immutable fastpath avoids
most of the work. This is still about half of the raw write performance,
so there may be more low hanging fruit that I missed. Its likely that
having to use a `compute` drops performance by losing optimizations
that can be done on a simpler `put`.