Add incremental reset option of sketch to the simulator
TinyLFU resets is a periodic sweep of halving all counters in the
sketch. An alternative is to reset incrementally, halving a table
location every N additions and incrementing the cursor. This would
remove a concern of an amortized O(1) cost being O(n) when on the
period to a fully O(1) operation.

This appears to work well for large caches with long traces. It has
a negative impact on small caches with short traces. Earlier ad hoc
experiments indicated this might be promising. Sadly that code was
not kept and the current analysis doesn't match those observations.

This requires more experimentation to see the feasibility of the
approach.