Incremental reset analysis
The periodic reset ages the sketch every 10 * max_size increments. For
Caffeine this is handled in the background maintance cycle, but others
may expose that penalty to a caller. Thus its a concern of a seldomly
seen latency spike, which is a type of hiccup that can be hard to
debug.

An incremental reset ages the sketch as a gradual sweep, making the
penalty constant at a slight cost to the accuracy. However, as we use
the filter to detect heavy hitters and SLRU to find promising victims,
the impact should be minimal. Here we reset an array slot (16 counters)
every 16 additions. This is done in batches of 16 for efficiency and
should be similar to doing 1 counter reset per addition.

The result is a range of +/- 0.5% across a variety of traces [1]. This
spikes to 5.5% for MergeS as the worst case. That data was taken before
realizing that no-op increments were counted for reset. With this change
that MergeS case dropped to a 3% penalty. I haven't refreshed the data to
account for that fix, but this is enough to draw an early conclusion on
being a reasonable design choice.

[1] https://docs.google.com/spreadsheets/d/1G3deNz1gJCoXBE2IuraUSwLE7H_EMn4Sn2GU0HTpI5Y/edit?usp=sharing