Add CaffeineSpec for externalized configurations (fixes #37)
Similar to Guava's CacheBuilderSpec, the CaffeineSpec simplifies the
creation of a builder from externalized flags. This can be useful for
configuration files, such as Spring properties.

The string format and options are the same as Guava's, except that
`concurrencyLevel` is not supported. The `disableCaching` factory
method was not ported over since its usefulness seemed questionable.

CaffeineSpec passes Guava's test cases, with the above changes made.