Document behavior of invalidating loading entries (fixes #13)
The behavior of `invalidate(key)`, `invalidateAll(keys)`, and
`invalidateAll()` is undefined for entries that have not materialized.
The undocumented behavior of ConcurrentHashMap is that a remove(k)
and clear() are blocking for loading entries. However the collection
views (keySet, values, entrySet) suppress loading entries. This can
result in the cache ignoring those entries, as it requires post
processing on removal.

As noted by Doug, the behavior is racy and difficut to document. The
expectation is that users arrange quiescent points are needed.