Prevent GC nepotism in write buffer's linked list
As described in [1], this scenario was not covered in the write buffer
(SingleConsumerQueue). This optimization only applies to the optimistic
queue (as used by the cache). The linearizable version cannot prevent
it due to needing to walk the chain to notify completion. It can reduce
it by observing that the node has been consumed and break the chain.

[1] http://psy-lob-saw.blogspot.com/2016/03/gc-nepotism-and-linked-queues.html