Ensure exception in maintenance task is logged
ForkJoinPool's internal runnable wrapper rethrows the exception, while
ForkJoinTask makes it available for the consumer as a Future result.
Adding this in as a safety net to let users know an unexpected failure
occurred.

I don't recall why exceptions from performCleanup were allowed to
propogate to the caller. This has something to do with the CacheWriter.
An explicit catch and log still abides by the old behavior, as we were
relying on FJP to log for us with the raw runnable.

Need to revisit and understand if propogation still makes sense.