CASSANDRA-11452: Improved resistence to hash collision attacks
When the candidate and victim's hash code are equal they have the same
estimated frequency. This results in the candidate being rejected. If
a hash collision attack then the cache will not admit any new entries
and the hit rate will be very low (causing poor response times).

Ideally we'd use a 64-bit hash, but Java does not provide that natively
and, for now, I'd prefer not exposing a builder function (K -> long). A
longHashCode() would not fix the problem, but greatly reduce its
likelihood.

This patch evicts the victim in the case of a collision. This should
allow candidates to flow through the eden -> probation segments. This
should have a similar effect as using freq1 >= freq2 comparision (not
strictly >) without degrading the hit rate in the LIRS traces.

@blambov please review so that we can iterate and find a good
compromise solution