CASSANDRA-11452: Stronger resistence to hash collision attacks
Thanks to further discussions with the Cassandra and TinyLFU folks,
this is a fairly strong guard against the collision attack. This also
has very low overhead and impact to existing hit rates. The previous
commit was fairly weak in comparison.

We concluded that some randomness was desirable in either the admission
or eviction policy. A few of the better options were simulated and this
is my favorite balance. It might change when Roy presents his feedback
from pondering it on his long flight home.

The attack requires that the victim be pinned to an artifically high
frequency due to a hash collision, so that no entries are admitted.
When a warm candidate is rejected by the frequency filter, we give it
a 1% chance of being accepted anyway. This kicks out the attacker's
victim and protects well against her having multiple collisions in her
arsenal. In small and large traces it is rarely triggered due to normal
usage, and does not admit too frequency to hurt the hit rate.

The other attack vector is the sketch's hash functions. This is already
protected with a random seed. We now protect against the weak 32-bit
hash codes used by Java.