Add write back-pressure for insert-heavy workloads
The write buffer is used primarily when a modification changes the size
of the cache (insert, remove, or update of an entry's weight). The
previous buffer was unbounded due to simplicity, lack of real-world
data indicating an issue, and stress tests showing a yield between
writes allowed the cache to keep up. While an unbounded queue is never
great, given the read heavy nature of cache workloads it hadn't been
a visible issue.

The bounded buffer grows from an initial size (16) to a maximum (based
on the number of cores). This allows a low footprint for ad hoc usage,
while also scaling well for high-demand usages. In a simple stress test
it achieved a good eviction rate (2.55M/s) with no GC penalty.

The implementation is a trimmed down version of MpscChunkedArrayQueue
from JCTools. Given the compatible license, friendly author, and great
quality code, it seemed poor to implement a custom variant.