Early pass at an adaptive TinyLFU
Gil's idea is that instead of using an admission window and adapting
its size, we adjust the emphasis given to a new arrival. In an LRU
workload this should adjust to favor the most recent and evict the
least recent. In a frequency workload it would behave like TinyLFU by
favoring the historic usage. The adjustment is determined by a sample
to observe if the policy has made good predictions.

This version shows uneven results. In some workloads it performs on
par with the adaptive W-TinyLFU. In others it is very poor. Tuning
the parameters causes the gains in one trace to hurt another. More
analysis is required.