Eclipse Neon import fixes
Prior to Neon, a Gradle import using Buildship was broken due to that
plugin not handling generated sources properly. It would force the
build directory to be marked as "derived" and not consumed by the IDE.
This was fixed in Neon (tested with RC2).

Cliff Click's highscalelib includes an optional trick to replace
ConcurrentHashMap with NonBlockingHashMap by using the bootclasspath.
If not used then the classloader should resolve to the JDK's, e.g.
rt.jar. In Neon it choses Cliff's instead and then breaks due to not
having atomic computes and lacking methods like `mappingCount`. The
likely culprit is Jigsaw changes. NBHM was moved into JCTools as it
was otherwise no longer maintained, where that trick was dropped. By
switching dependencies this build issue goes away.

Lastly Eclipse now makes using Unsafe an error. We're waiting for JDK9
VarHandles, so premature. That requires manually lowering the threshold
to a warning.