High false positive rate in simulator's BloomFilter (fixes #85)
See the issue for the discussion. Clearly I goofed in my thought process
and should have taken the highest most bits. This had no effect during
experiments due to the very small BF used during adaption, which is why
I didn't see a difference compared to Guava's. The issue would have been
a problem with the doorkeeper when that idea is analyzed.