Fix jsr107 CacheLoader adapter to handle null values (fixes #119)
This is required by the specification (per JavaDoc) but the TCK does
not validate it, so sadly I failed to handle that case.

To get CI functioning again, moving back to a jdk8-only config. I don't
know why the new config hangs with Guava's tests and think it might be
the distro.

Also fixed the misnamed method for DI factories. I'm going to argue it
doesn't break semver since JCache is the API and this is an impl hook
no one calls directly.