Update shaded JCTools queue due to live-lock (fixes #127)
In #127, a live-lock occurs in the write buffer due to a race when the
consumer tries to transition to the last buffer. This is a shaded copy
of JCTools' MpscGrowableLinkedArrayQueue and fixed in 2.0 (see
https://github.com/JCTools/JCTools/issues/135). This is a difficult to
reproduce bug, except for the reporter due to having thousands of
instances on a loaded system.

The shaded copy is more verbose than the previous one, but that is to
minimize the divergency and introducing bugs by cutting too deep.

Updating to Guava 20 required minor changes due to that cache now
supporting write-through entries. The adapter now mirrors the new
behavior.

Thanks to everyones patience and help in fixing the bug.