Add Coverity build scan
Fixed minor compilation errors in their static analysis. But the plugin
assumes a commercial installation for local analysis. The open-source
portal doesn't seem to accept the emit-db due to not having a
build-log.txt file. So more debugging is required to figure out how to
integrate with Coverity Scan. Despite that, merging since the changes
were benign and we might get it working in the future.