Fix stale notification when clear raced with an update
Similar to 83b47d1, clear did not guard against a concurrent update and
ensure that the notifications were not given a stale value. A version
of the Solr test case demonstrated this bug occuring and that this
change resolves it.

Added two unit tests that rely on the synchronizations to demonstrate
ordered blocking and that the latest value is notified with. This
depends on the changes and cannot be used as a generic sanity test
like Solr's.