Some variable expiration unit tests
The expiration time is now set after the writer has been called. The
evaluation is done prior, so that an exception thrown by expiry must
occur before the writer is notified (as the operation is rolled back).
The entry's expiration time is set after, so that a writer exception
does not extend the entry's lifetime.

A stub of the inspection interface was defined, but not yet implemented.

A few unit tests were added, though more are needed for this feature.