Added ability to inspect the variable expiration policy
Supporting a snapshot view of the entries in expiration order required
adding an overflow bucket. Previously overflow was scheduled on the
furthest away bucket, but this meant order could be less exact when
traversing. Note that order within a bucket is not computed so the
view is only a best guess.