Integrate variable expiration with JCache
JCache has its own expiration policy, but unfortunately differences in
the API and the requirement to retain existing functionality means that
the two cannot be merged yet.

The settings currently allow using both variable and fixed expiration
together. This is because the spec authors assumed that a maximum size
would also be enabled, thereby evicting expired entries lazily. Since
we exposed Caffeine's eager policy, then only for fixed, these two
could be used together. With Caffeine's new variable policy, it does
not allow the combination of fixed and variable in the core library.
Thus, lazy-expiration cannot be mapped directly to Caffeine's version.

To retain the previous combinations, we could adapt only when the
fixed settings are not used. Then supply an eternal no-op policy for
the lazy logic and an adapter for the eager logic. Unfortunately API
differences makes this more difficult. For example, JCache's
putIfAbsent does not return a value when failing and therefore is not
a read. Caffeine's does, which results in TCK failures when the expiry
policy calls are verified.

A possible solution would be to use a ThreadLocal to signal the cases
when the expiry calls should be ignored. That adds brittleness to an
already complex adapter. Instead, perhaps in v3 we can remove the lazy
and fixed settings, simplify the adapter, and use the ThreadLocal
trick to correct these edge cases.

Given that the major JCache implementors are lazy, expect a maximum,
and that JCache's ExpiryPolicy is neutered by not providing the key or
value, I don't think a hacky solution is worth the effort. Users can
use Caffeine's expiry if they need eager behavior.