Honor unsetting JCache configuration value + maximum size tests
When a value in the default configuration is set and the cache unsets
it (`value = null`), the merged result is the default's. To honor the
desired behavior additional inspection is needed of the different
configuration layers.

Added unit tests to assert that a maximum size eviction is honored
and that the removal notifications are published. For simpler tests
the executor is now a configuration option to disable async processing.

This uncovered that a custom Weigher or Expiry was given the unwrapped
value which would cause a ClassCastException. This is now delegated
properly.