Fix TravisCI build
Due to exceeding the 50 minute time limit, the support team was very
kind by extending it to 90 minutes. For now, merging this slow tests
into the same job for a fuller code coverage report.

Moving to a VM instance (vs container) for a beefer machine. This had
hung because some Guava tests block indefinitely on CountDownLatches.
Since I had switched their tests to use a pool instead of new threads,
thread starvation meant no progress could be made. Reverted to ad hoc
threads and added a timeout to the awaits().