Fix null mappings caused by JDK-8186171 (fixes #176)
When a getAll or getAllPresent is called, the duplicate keys are
discarded and the lookups performed. Previously this was done by
building a mapping with sentinel values, iterating the entries,
performing each load, and setting the value or removing accordingly.
That approach avoided garbage of a unique key set and let the hashmap
perform faster set and remove operations.

Unfortunately this was broken in JDK8 when a bucket was treeified due
to collisions and untreeified by the removals. That caused the link
chain to be rebuilt and the entries decoupled from the map. Then the
sentinel value leaked into the result set given to the user. We used
null for convenience. This was a regression from JDK7 and a bug.

The more obvious and slightly less efficient version is now implemented.
The keys are stored copied into a HashSet and the map populated only
when a value exists. This should have similar performance in practice.

https://bugs.openjdk.java.net/browse/JDK-8186171