Wait on future outside of a computation in Async's asMap sync view
The AsyncLoadingCache's synchronous asMap() view has to block on the
existing in-flight future when performing a computation, such as
computeIfPresent. That blocks other computations for the same key,
such as eviction's computeIfPresent to remove the entry. If the
future itself is a cache write, then the recursion may hit the
ConcurrentHashMap live-lock.

Instead the entry's future is retreived and blocked on outside of
the computation, validated as the same instance inside of the block,
and the mapping function executed or else the entire operation is
retried.