Do not propograte null value in completed future (fixes #206)
AsyncLoadingCache stores futures as the cache value when they have
completed successfully and are not a null result. While the entry
was being properly discarded, the null result could be propogated
to the callback evaluators (weigher, removal listener, and expiry).
While null handling was tested, the various implementations did not
check the parameters before returning their fake response.