TinyLFU paper was published by ACM's Transaction on Storage!
This was a long, two year process from the pre-print to the finalized paper. The
major differences are clarity, fixed an incorrect chart for OLTP (was accidentally
TinyLFU+LFU, much worse than the reported TinyLFU+LRU), and appendix for reviewer.

This link is from ACM's authorizer service to comply with copyright requirements.
This works from the project page, but will not work if copied elsewhere.