Add NullAway for more extensive static analysis
NullAway is an addition to ErrorProne to perform a more robust data
flow analysis for possible NPEs. This inferrence requires richer usage
of the defacto null checking annotations by defaulting to a non-null
expectation.

Due to limitations in the data flow, there are false positives which
requires slightly more verbose code or suppression. When possible,
the suppressions are isolated to individual statements to minimize
their scope.