Added a compute time benchmark based on the traces
The goal of this benchmark is to simulate the performance benefit of
switching algorithms on a fixed workload. If the hit rate increases,
one would expect to see the execution time decrease as the cost of a
miss goes up. However, if the eviction policy's overhead outweighs
the workload's costs then the optimization is not worthwhile.

Therefore the goal is to show if the improvements by a more advanced
algorithm outweigh its costs. For example LRU is extremely cheap to
execute and, with a small penalty, TinyLFU can be added to increase
its hit rate. In a workload like ARC's DS1 trace (database), this is
clear cut. But if we introduced a new policy that was very expensive
to operate, it wouldn't be feasible in practice.

This isn't expected to be used as a microbenchmark to performance tune
with. Rather its an indicator of whether a technique, such as mini-sim
sampling, is feasible in practice.