Link to our new research paper on adaptive eviction policy
Adaptive Software Cache Management, ACM Middleware 2018
https://dl.acm.org/citation.cfm?id=3274816

The paper can be freely downloaded by following the link in the README
if coming from this github. This uses ACM's Authorizer service.

Congrats to @ohadeytan for this work winning Middleware 2018's best
student paper award!

We plan on integrating ideas from this work into a future release of
Caffeine. The research source code is available in the simulator.