Minor cleanup of comments and naming
The comment separator was also made balanced, and amusingly was copied
from ConcurrentHashMap where the typos remain.

The very early name "eden" was replaced with "window", the name we chose
for W-TinyLfu but hadn't updated the code to. That leaked when people
read the code and started using the initial term.

This splits most of the verbose, unimportant changes from the next
commit to keep each more focused. For simplicity that commit contains
the related cleanups in addition to its improvements to the affected
files.