Async map support (fixes #156)
AsyncCache now includes an `asMap()` view where the value is the
`CompletableFuture`. This allows for directly working with the
futures with all of the flexibility that Map provides, just as is
possible with the synchronous cache. The synchronous view has the
limitation of requiring that the existing future completes and the
new value cannot be in-flight.

Because the future (put, replace, computes) may be in-flight, the
completion of the future is counted as the load success / failure. The
map operation to insert that future is not counted in the statistics, as
typically the future is immediately obtainable, fetching it shouldn't
fail, and the success / fail / cost is not a meaningful attribute. This
means some operations record load stats differently than the synchronous
view. The topic of stats is fuzzy enough that this seems okay, as long
as it provides some meaningful metrics that fit the mental model of the
cache type being used. A related difference is that futures do not have
a notion of equality (reference-based) which can cause subtle differences
in behavior.