Add removal cause to StatsCounter (fixes #304)
This provides more details to the collector, which may optionally use
it. This differs from the previous norm of collecting all details into
a default implementation and providing them through CacheStats.Here and
possibly future cases, CacheStats and the default in-memory collector
won't be expected to provide this level of granularity. Those that want
more details are able to provide their own StatsCounter and handle
appropriately.

Additional counter methods will be added as requested, typically with no
corresponding implementation for the built-in support.