Migrate to the maven-publish plugin
Previously releases were performed using the legacy 'maven' plugin. This
is hard to configure, so we used the 'nexus' plugin built on top of that.
Since then, that plugin inspired the official maven-publish plugin which
uses a similar configuration. In Gradle 6 the legacy plugin was deprecated.

The output is similar except it now includes a more succinct generated pom
(no test scope dependencies) and the new Gradle module metadata.

The simulator's application archives (zip and tar) are no longer attached,
as they were by default before. Since this is a development tool and there
isn't a known use-case for publishing those, until otherwise asked for they
are no longer provided.