Support simulations of large caches in weighted policy types
In weight-based traces the entries have different sizes, e.g. bytes. The
policies that support this characteristic will evict multiple entries
until at its maximum-size contraint. This was an integer setting which
is too small for the Tencent Photo trace.

The setting was increased to a long and the weighted traces updated
accordingly. The other traces use a checked cast to an int value as a
minimal migration to minimize the chance of introducing bugs. The
linked policy was fixed to handle the entry's weight changing on a
cache hit.

The AdaptSize simulator compatibility was reduced to only its trace
reader. That simulator treats the same key with different weights as
distinct entries, rather than an update. The workaround of hashing
the key and weight as the entry's key causes noticable workload skew.
Since other simulators (like LHD from the same group) don't make this
odd choice, its now restricted to just their trace reader.