Migrate to CLA Assistant; fix flaky test
CLA Hub silently disappeared, along with the signatures it gathered.
Thankfully the Apache License covers all cases, such as contributions,
so this was not required. We offer a CLA in case if alleves concerns
by a corporate legal review. The same CLA terms apply: a direct copy
of Apache's with the names updated.

A test appears flaky on the CI due to weak/soft values being collected
during its execution. The test now better ensures that the entries are
not collected, hopefully stabilizing this test case.