Allow for simulator policies to add custom metrics
In addition to the common policy statistics, like the hit rate, there
may be a desire to track algorithm-specific metrics. For example a
Clock policy will perform an O(n) scan on eviction to unmark hot
entries as it searches for a cold entry to evict. It might be useful
to capture the worst case run length observed for a given trace.

A policy can extend the PolicyStats to register and track new metrics,
and the reporter will capture them in its output. For efficiency the
metric are registered by a supplier, rather updating a mapping to the
value itself. This avoids a high rate of object churn if modified on
every access event (which may be millions). Instead the metric should
be kept as a primitive instance field and the accessor method
registered.

If a metric is not captured by a policy then the reporters print
an empty result. If nothing was captured across all policies, then
the metric is dropped from the report to reduce visual noise. This
also includes metrics that are tagged by a feature characteristic
which was not employed by the trace (such as entry weights).