Improved simulator runtime for large traces
Unfortunately, the elegant approach of partitioning a stream of events
into batches using the Iterator view comes with a hidden performance
cost. Instead of picking off events one-by-one, large internal buffers
are created and grown (SpinedBuffer) before the first element is
provided to the consumer. The maximum chunk size is 1M elements, which
can cause a long pause while the trace is read into memory before it
is processed by the simulator. For the DS1 trace this adds 10 seconds
of startup time for the simulator and doubles the time for the trace
rewriter.

The forEach operation pulls on-demand, where our smaller I/O buffers
and concurrent prefetching keep the policies fed with a lower startup
overhead. This also greatly reduces the memory footprint to allow for
more easily analyzing large caches. However this coding structure is
slightly less pleasant.