fix simulator's bloom filter table minumum sizing
The table index is calculated by shifting the hashed value. As no
modulus is applied, the negation bit may lead the minimum index
to be 0 or 1, while the table size is one. Either the index has
to be modulated (via a mask) or the smallest table should be 2
to avoid this off-by-one error. Increasing the minimum allows for
simplifying the code.