Fix NPE when on the bootclasspath (fixes #481)
To reduce runtime overhead due to fields needed only by certain
configurations, code generation is used and reflection loads the
optimal implementation classes. This trades off a small amount of
disk space for reduced memory usage.

Unfortunately the wrong classloader method was used, which is not
available on the bootclasspath (null value). This fails when the
cache is used by a compiler, e.g. ErrorProne's javac fork. Instead
Class.forName is now used which should resolve to a non-null,
acceptable class loader. It appears as if this should be the same
loader as before, or else the system loader if on the bootclasspath.