Lazily initialize simulator policies
Previously the metadata for what characteristics a cache policy
supports was on an instance method. This required creating the policy
and then filtering out those that do not support the required features.
This is problematic because a policy's constructor perform allocations,
such as initializing arrays based on the maximum size. For large
weight-based runs, this could allocate very large objects that were
immediately thrown away because the policy is not compatible with the
configuration constraints.

All customizations for the metadata is now defined on an optional
@PolicySpec annotation. This is inspected and filtered against,
thereby avoiding the unnecessary creation.

This cut the execution time of a 1gb ObjectStore trace from 1.5m to 50s.