Prefer tighter generic bounds on query methods
The Collection Famework's idiom is to be strict on methods that insert
and loose on those that query. For example add(E) and contains(Object).
Guava's cache followed this practice and relied on static analysis to
discover poor usages when the type would never match. That put the
burden on developer tooling for flexibility that isn't very desirable.
See [1] for the technical details behind the idiom.

We now use stricter bounds in hopes this is a friendler experience.
Thanks to asMap().get(Object), there is an escape hatch anyway.

[1] https://errorprone.info/bugpattern/CollectionIncompatibleType#footnote-would-requiring-e-have-been-better