Introduce eviction listener & deprecate CacheWriter
A RemovalListener may now be configured that is called atomically during
an entry's eviction. This listener will be invoked on the calling thread
during the map computation, and only be notified when the cause's
wasEvicted is true. Any thrown exception will be logged and swallowed.

This replaces the most common case for CacheWriter. The other cases of
atomically handling explicit writes and deletes can be better handled
by the various asMap().compute methods. These methods offer more
flexibility and the behavior is clearer.

This listener may not be used when both weakKeys and AsyncCache are
used together, as the future may be evicted prior to the value
materializing. CacheWriter is incompatible when either is set.

CacheWriter is deprecated and scheduled for removal in 3.0.