Eagerly evict an entry if it too large to fit in the cache
When an entry's weighted size exceeds the cache's maximum size, then
it can be evicted immediately and avoid flushing existing entries.
Some protection was already built in, but it was not robust to cover
all edge cases. Thus, it could flush some entries such as the entire
admission window on insert, or many others on an update.

When the entry is being inserted or updated in the admission window
then it is bounded by the window's maximum size (e.g. 1% of total).
If so, then the entry is placed immediately at the window's LRU
position so that it is promoted into the main space and treated as
an eviction candidate.

When the entry resides in the main space then it is bounded by the
cache's maximum size. It is moved to the window's LRU position so
that it becomes the eviction candidate.

This improved protection can cause a weighted cache to  discard the
most recently added entry. We typically try to minimize that to
avoid surprises, as one generally assumes a different victim will
be chosen. On a subsequent add the entry may be retained, thanks
to building up its frequency. Therefore, the preference is to
protect the cache at the cost of breaking an invalid assumption
for those step debugging through a sequence of operations.