Add restrictions for when eviction listener can be used
An eviction listener cannot be combined with weak keys and an async
cache. This is because the entry can be evicted when the key is
collected, but the future is still in-flight. As the listener is
atomic with the map operation to remove the entry, this would
require blocking and is therefore not allowed.