Add support for releasing by github action
Uses the nexus-publish plugin to close and release on the Sonatype
repository. The GPG key is registered by a secret for signing. This
workflow is conditioned to creating a new release tag.

The intent is to allow anyone with commit access to perform the
release. This makes the project more resiliant than being dependent on
a single developer. It also streamlines the process by automating it.