Added snapshot view of in-flight refresh operations
The ongoing refresh operations can now be observed by using
Policy.refreshes(). This does not provide a live view due to weak
keys needing to be unwrapped from reference key (and possibly
discarded if null). Instead an unmodifiable copy is provided for
inspection. The futures may be canceled, completed, or waited upon.