Remove sun.misc.Unsafe usages (was optional)
Previously Unsafe was used if available for appending into the read and
write ring buffers. If unavailable then a fallback was used, such as
VarHandles.

This was most valuable for the read buffers which is a hot path. The
previous commit removed that case by reworking the probing strategy.

The write buffer is a less critical path and can use VarHandles
for similar performance.