Remove non-null type argument (fixes #530)
Checkstyle interprets this as defaulting the type to non-null, but will
also do that implicitly on a raw `extends Object`.

IntelliJ Idea interprets the generic type annotation as requiring that
the type itself is non-nullable, e.g. `String` versus `String?`. This
cannot be expressed in Java, so the warning is impossible to resolve
and can only be suppressed.

As the type argument adds no additional benefit for Checkstyles,
removing may make the signature look odd but better declairs the
intent.