Systor17 trace reader cleanups
This trace is for a block device where every block is 512 bytes. Since
the blocks are uniform, a disk cache would store them by blocks instead
of by object. Therefore mirror the UMass and ARC trace readers by
generating the block sequence. This increases the cachability and the
number of events in the trace file to match what the system is
experiencing.

The systor distribution is a tar file with many gzipped files. Previously
it was assumed a tar.gz with uncompressed entries. The trace reader will
now recurse to handle compressed entries.

Minor fixes empty event streams and disabling jigsaw modules as Eclipse
does not handle them well yet.