Fix timer wheel cascading and support cancelating a scheduled task
When the timer wheel advances then higher level buckets evict to the
lower wheels. In some cases this reorganization failed, such as when
the nano time transitioned from a negative to a postive clock value.
The number of steps to turn the wheel is now determined by a more
accurate calculation. (fixes #541)

When a scheduler is enabled then a future task may exist against the
next expiration event to perform cache maintenance. If the cache
becomes empty beforehand, e.g. Map.clear(), then this future can be
canceled. (fixes #542)