fix compilation on Windows machines
The excludePaths option for ErrorProne is used to ignore warnings in
generated code. That supplied the path using `buildDir`, but in fact
this is documented as a regular expression. On Windows the separator \
is an invalid regex character if not escaped. By instead using a more
generic pattern, rather than a file path, the build is cross-platform.

https://github.com/google/error-prone/issues/2394