Use custom Integer type for unit tests
The cache unit tests used Integer as the key and value type, which
allowed for simpler code thanks to autoboxing and the JLS int cache.
However, some unit tests require identity semantics such as if weak
or soft reference caching is used. In such cases the Integer(i)
constructor was used.

Java 9 deprecated the constructors for removal so that a future release
can make the class a value type. To migrate, a custom Int(i) type is
introduced that mirrors Integer and the tests use this instead. Similar
to Integer, valueOf(i) may return a cached instance if within a known
range. The type also includes many convenvience factory functions to
restore the succinct code that autoboxing afforded, such as when
creating lists and maps.