Run Guava's conformance tests across more configurations
The core test suite is parameterized so that the scenario it validated
for every applicable cache configuration. This allows us to avoid
mistakes that only apply when a feature is enabled, and could otherwise
be missed due to a test applying only the happy path configuration.
Guava's testlib provides a test generator for the Collection Framework
so that implementations of those interfaces can be verified generically.
Previously only the base configurations were tested, but now many more
are verified (380 configurations, resulting in 350,000 scenarios).