Increase performance and test coverage
In the JMH read benchmark, swiching to the weaker form of compareAndSet
can increase the throughput. This is used when recording a read event in
a ring buffer, which is highly speculative. As these events are allowed
to be lossy, allowing spurious failures due to memory contention by hot
CASes has no negative impact. This reduces the strain on the hardware's
cache coherency by failing faster, which increases throughput under
high load.