Suppress log warning if a future is cancelled or times out (fixes #597)
Caffeine logs a warning with the exception when an asynchronous load or
refresh fails. This is to assist in debugging due to a common mistake
of not handling errors, e.g. chaining `thenApply` without an exception
handler in a fire-and-forget fashion. It is expected that if this aid
is an annoyance that application can simply disable this in their
logging configuration.

This logging is not helpful when the future fails due to calling
the future's cancel() method or using orTimeout(duration). In those
cases the cause is expected and desired, so we can assume that the
application is handling the failure appropriately. Therefore the
logging of these causes should be skipped.