Add tests for early bulk completion
When a bulk async load finishes it has to complete the placeholder futures
that are present in the cache. In a pre-release change, this was switched
from the obtrude methods to a soft complete due to forgotting the rational.
In revisiting it for tests and asserting the expected behavior, the
original approach was correct and is restored.

Any early completion of an in-flight future by a non-loading thread will
violate all assumptions that one can make. When the load completes its
results will be silently discarded (no removal notification), while if
the future was done then the new result will not invoke the callbacks (so
no automatic removal if exceptional). The bulk completer can only post-
process the results correctly if it forces the future into the desired
end result, so that the cache is back to a consistent state. While the
obtrude methods appear harsh, it is only in reaction to bad behavior if
users intentionally cause the future to resolve in an unhandable manner.