Detect JDK-8274349 and fallback to a same-thread executor
In certain scenarios the common pool is broken and will never execute
a subitted task. This causes behavior like async caches, evictions,
removal notifications, and refresh after write to not function. The
application's own usage will not work either so this only fails less.

Affected users are encouraged to use one of the following workarounds:
- set the property `java.util.concurrent.ForkJoinPool.common.parallelism`
  to 1
- do not use `-XX:ActiveProcessorCount=1` as a startup flag
- do not use a container with 1 cpu

For more details see:
  ForkJoinPool.commonPool() does not work with 1 CPU
  https://bugs.openjdk.java.net/browse/JDK-8274349