Remove detection of JDK-8274349
In retrospect, while a nasty bug, it is not the library's responsibility
to detect and quietly fallback. The application would still be broken in
all other usages of the FJP commonPool. We should only sympathize with
affected users and inform them of a likely culprit. Ideally their system
heath tools would be able to detect probematic configurations like this.