The bulk loader should not be allowed to modify the keys to load
Previously the modifiable set of keys to be loaded was given to the
loading function and then iterated for the results. This could cause
non-deterministic behavior if the user added or removed keys, as it
was assumed to be immutable. This enforces that implied contract. The
same oversight was found in Guava, google/guava#5810