Allow for more advanced snapshots of entry metadata
Previously the entry metadata had to be queried one-by-one, which could be
too limiting. The Policy.CacheEntry interface allows for reading multiple
peices at once, where getEntryIfPresentQuietly(key) offers a single lookup.

The ordered snapshots (hottest/coldest, youngest/oldest) now allow for
obtaining these entry types. To support cases where the limit is not
known ahead of time, e.g. to skip over some entries, the snapshot
provides a consumable stream. The computations should be short and simple,
prefering to produce a result for downstream processing, to minimize the
hold time of the eviction policy's lock. Therefore this api should be
used with care and is intended for advanced integration scenarios.