Add ability to specify the expiration time with a compute
If variable expiration is enabled, one can pass in a a duration
explicitly and override the configured Expiry. This is useful
for integrations to conform to an existing api, e.g. a put
that has synchronous callbacks for create and update.

This variant is still left limited by not allowing for dynamically
determining the duration as part of the compute's atomic operation.
That is left for a future if needed, as it adds API bloat without
a reasonable use-case (e.g. integration with a legacy api). It is
preferred and less error prone to use the Expiry, so we'll keep it
straightforward for now and add later if user requested.