fix flaky assertions for reference caching tests
The tests rely on GC completing, clearing the references, and the cache
able to discard them by polling the reference queue. As the GC runs
asynchronously sometimes a race will occur that causes the tests to
think the cache size has stabilized when it has not. Instead pass a
target size to continuously clean up until that is reached or the test
runs out of attempts.

Other changes are simply for improved readability and strengthening of
the assertions.