Add tests for thread interruption
When a cache is interrupted by the CacheLoader, it sets the thread flag
and wraps the InterruptedException for a runtime throw. This matches
Guava's behavior so that status is communicated without having the
checked exception in the interface contract. Neither cache apply this
rule when the mapping function is provided since the caller had to
handle the interruption themselves to bubble it up as a runtime
failure.

This applies to an async cache if interrupted while obtaining the
future. If the thread computing the future is interrupted then that
is merely the failure reason, but the consumer is not interrupted.

Interruption can occur silently when performing a refreshAfterWrite's
reload. This too sets the flag, but does not throw an exception as
documented that reload failures are logged and suppressed. The caller
can still check its interruption status, e.g. to halt its processing
in classic threaded logic.

Generally we don't recommend using interruption as a means of
communicating cancelation as a confusing, error prone, and not widely
used mechanism. Likely virtual threads will promote a better approach,
or at least evangulize interruption so that developers understand how
to use it properly.